Bump node (20.14.0 to 20.15.0) & specify node version in .nvmrc file (#218)
Motivation for using a `.nvmrc` file: I noticed that `yarn install
--check-files` was failing on my machine. This was because I was
actually running node 20.8.0, whereas a package required at least node
20.9.0. I was running 20.8.0 despite the fact that our GitHub Action is
running node 20.14.0 (and so was succeeding, even though I had a problem
locally).

This ability for the local environment to be out of sync with the
CI/deploy environment is not good. This change (to use a `.nvmrc` file
as the source of truth for both CI/deploys and also locally) should
prevent such misalignment in the future.