Fetch full git history in GitHub Action (for accurate timestamps)
The `bridgetown-sitemap` gem uses the git history to determine the last
modified time of each post (if one is not manually specified in the
post's frontmatter).

Prior to this change, we were only fetching the latest commit, and so we
were not able to determine (using git) an accurate `lastmod` time for
each of the sitemap entries.

With this change, we will now fetch the full git history, which should
allow `bridgetown-sitemap` to correctly determine the time at which each
article was last modified.