[seo-tags] Remove double '/blog' path fragments
I'm not sure if this is a bug in `bridgetown-seo-tag`, or something that
I am doing wrong, but, whatever the cause, in production, we currently
have two URL references in the SEO tags on each blog show page that are
like
https://davidrunger.com/blog/blog/mocking-external-requests-in-rails-feature-tests
with `/blog/blog`. This change condenses that down to just one `/blog`.

This is hacky, but I think it's fine. There's probably a better way to
fix this (maybe an upstream bugfix), but I'm just going to do it this
way, for now.