Change githooks link from atlassian.com to githooks.com (#454)
The Atlassian URL sometimes generates emails to me like this (via the
nightly davidrunger/david_runger link checking worker(s)):

> We made a request to
https://www.atlassian.com/git/tutorials/git-hooks, which is linked from
https://davidrunger.com/blog/enforce-git-hooks-in-a-rails-initializer,
and its response status was nil, but we expected it to be in [200].

To avoid these emails, I'm changing the link to githooks.com, which will
hopefully be more reliable.

Also, I am updating the quote to reflect the new source.