[posts/rails-app-update] Use $(pwd) (not hardcoded absolute path) (#506)
I got this idea by noticing that a ChatGPT user agent was visiting this
blog article. That made me curious to see whether and how ChatGPT would
reference the article in a chat. I prompted: "How can I use VS Code as a
merge tool for Rails's `app:update` command?", and it replied largely
with a summary of my article, but it included this slick and very
sensible change, which allows me to avoid the somewhat awkward mention
about how my absolute path will differ from the reader's and which also
makes it more convenient for the reader to simply copy-paste the
command.