Use Percy for visual testing [BLOG-21] (#559)
Hopefully this will catch any regressions that might be introduced by a
Tailwind update or by my own code changes, for example.

It was also useful to realize that the blog has a light mode, in
addition to the dark mode that I always view it in...

This also showed me some visual bugs with the "Copy Code" button (on
small screens / where it renders over the code) that I had overlooked
before.