[posts/mocking-requests] Update Mozilla CSP link (#562)
Email that I received today:

> Link to https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP did not
return expected status
>
> We made a request to
https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP, which is linked
from
https://davidrunger.com/blog/mocking-external-requests-in-rails-feature-tests,
and its response status was 301, but we expected it to be in [200].

This change should fix that.

Also, remove hard wrapping of the paragraph. I think I am going to stop
hard wrapping paragraphs in the blog and rely on developers and their
tools to handle that for them, as needed. The advantage of this is that
it will eliminate some unnecessary git diff noise when changing the
length of some text relatively early in a paragraph that affects how the
rest of the paragraph wraps.