Change link text for GitHub editing to "Improve this page"
I saw that here https://tools.simonwillison.net/ and I think it's a nice
improvement over "Edit this page". It provides a more compelling reason
for the reader to act, and it has a more positive and progressive feel
to it.

Also say "This blog is open source."

Also, move this stuff to the bottom of the page, rather than cluttering
the top with the date and author info before the user has yet even had a
chance to read the article.