Make links blue in light mode and code greenish and underline links (#580)
The goal is to make links look like links, and to make code not look
like links.

I sort of visually like the blue code that we had before (especially in
dark mode), but it's annoying that those code sections feel like links.
In that sense, I think that this change is an improvement.

### Dark mode before


![image](https://github.com/user-attachments/assets/22032804-4c5e-47f6-8126-c553331270f7)

### Dark mode after


![image](https://github.com/user-attachments/assets/d652fea4-fffd-4b49-9d1c-bb8f69f5f710)

### Light mode before and after


![image](https://github.com/user-attachments/assets/f978fcde-9ffa-4fa6-baba-5cbddc640ee1)