Switch to GitHub-style copy-code icon [BLOG-26] (#582)
I got the icon from here: https://icons.getbootstrap.com/icons/copy/ .

I think that the GitHub-style icon will be more recognizable to users
than the clipboard-style icon that we had before.

Also, the SVG is smaller.

Also, this change drops the button background (when not hovering).

Overall, with the icon change and dropping the non-hover background, the
icon looks cleaner now. I'm happy with it.


![image](https://github.com/user-attachments/assets/a6bfb9a4-325d-4336-8b0b-8360b2ae7ee8)