[footer] Remove year from copyright notice [BLOG-27]
I read on Reddit*, I think, that the year is not supposed to be the
current year, but rather the year that the material was created and/or
updated or something? Anyway, let's just remove it. I highly doubt that
it's ever going to really matter to me or anyone else whether I have the
year in the copyright notice or not.

*a reliable source of legal advice :-p

Also, move the copyright notice below the "View source at GitHub" link.
I think that the copyright notice is relatively less important to me and
to readers than the GitHub link, so I think it should appear later.