Specify 'margin: 0 auto' for post body using CSS class (#598)
The `.post` class will be applied to `body` for posts, via
`resource.data.layout` in `src/_layouts/default.erb`. We are specifying
this CSS here via a class to override Tailwind reset for `body` that
gives `margin: 0`. As part of
https://davidrunger.atlassian.net/browse/COM- 9 , we will start pulling
in the Tailwind reset (and Tailwind styles) when rendering the blog via
`david_runger`.